requirements.txt
moviepy
openai
Flask
requests
pillow
opencv-python